--- api-gateway/src/main/java/com/akentech/shared/models/ApiGatewayServiceApplication.java
@@ -1,11 +1,0 @@
-package com.akentech.shared.models;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiGatewayServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ApiGatewayServiceApplication.class, args);
-    }
-}
