+++ expense-service/src/main/java/com/akentech/kbf/expense/ExpenseServiceApplication.java
@@ -1,7 +1,8 @@
-package com.akentech.shared.models.expense;
+package com.akentech.kbf.expense;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 public class ExpenseServiceApplication {
