+++ transaction-service/src/main/java/com/akentech/kbf/transaction/TransactionServiceApplication.java
@@ -1,11 +1,10 @@
-package com.akentech.shared.models.income;
+package com.akentech.kbf.transaction;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
 @SpringBootApplication
-public class IncomeServiceApplication {
+public class TransactionServiceApplication {
     public static void main(String[] args) {
-        SpringApplication.run(IncomeServiceApplication.class, args);
+        SpringApplication.run(TransactionServiceApplication.class, args);
     }
 }
